requirements.txt
paho-mqtt
zeroconf
